Load pytorch from pytorch-nightly, not from conda-forge
The order of the channels list is significant. The package is installed from the first channel where it is found. Since we want to install the nightly version of pytorch, it should appear in the list first.

See https://docs.conda.io/projects/conda/en/latest/user-guide/tasks/manage-channels.html for exact rules of priority.